'AnalysisProject' and 'AnalysisSample' classes: remove QC-related methods which depend on qc.reporting

+++ auto_process_ngs/test/commands/test_import_project.py
@@ -7,8 +7,14 @@ import tempfile
 import shutil
 import os
 from auto_process_ngs.mock import MockAnalysisDirFactory
+from auto_process_ngs.mock import UpdateAnalysisProject
+from auto_process_ngs.mock import MockMultiQC
 from auto_process_ngs.auto_processor import AutoProcess
+from auto_process_ngs.analysis import AnalysisProject
 from auto_process_ngs.commands.import_project_cmd import import_project
+
+# Set to False to keep test output dirs
+REMOVE_TEST_OUTPUTS = True
 
 class TestAutoProcessImportProject(unittest.TestCase):
     """Tests for AutoProcess.import_project
@@ -16,6 +22,11 @@ class TestAutoProcessImportProject(unittest.TestCase):
     """
     def setUp(self):
         self.dirn = tempfile.mkdtemp(suffix='TestAutoProcessImportProject')
+        # Create a temp 'bin' dir
+        self.bin = os.path.join(self.dirn,"bin")
+        os.mkdir(self.bin)
+        # Store original PATH
+        self.path = os.environ['PATH']
         # Make a mock project
         project_dir = os.path.join(self.dirn,'NewProj')
         os.mkdir(project_dir)
@@ -37,8 +48,11 @@ Comments\t1% PhiX spike in
         self.new_project_dir = project_dir
 
     def tearDown(self):
+        # Restore PATH
+        os.environ['PATH'] = self.path
         # Remove the temporary test directory
-        shutil.rmtree(self.dirn)
+        if REMOVE_TEST_OUTPUTS:
+            shutil.rmtree(self.dirn)
 
     def test_import_project(self):
         """import_project: check project is imported
@@ -70,6 +84,51 @@ Comments\t1% PhiX spike in
         self.assertTrue('NewProj' in [p.name
                                       for p in ap2.get_analysis_projects_from_dirs()])
         self.assertTrue(os.path.exists(os.path.join(ap2.analysis_dir,'NewProj')))
+
+    def test_import_project_with_qc(self):
+        """import_project: check project with QC outputs is imported
+        """
+        # Make mock multiqc
+        MockMultiQC.create(os.path.join(self.bin,"multiqc"))
+        os.environ['PATH'] = "%s:%s" % (self.bin,os.environ['PATH'])
+        # Make an auto-process directory
+        mockdir = MockAnalysisDirFactory.bcl2fastq2(
+            '160621_M00879_0087_000000000-AGEW9',
+            'miseq',
+            top_dir=self.dirn)
+        mockdir.create()
+        # Add  QC outputs to the project to be imported
+        UpdateAnalysisProject(
+            AnalysisProject('NewProj',self.new_project_dir)).add_qc_outputs(
+                include_multiqc=False)
+        # Check that the project is not currently present
+        ap = AutoProcess(mockdir.dirn)
+        self.assertFalse('NewProj' in [p.name
+                                       for p in ap.get_analysis_projects()])
+        self.assertFalse('NewProj' in [p.name
+                                       for p in ap.get_analysis_projects_from_dirs()])
+        self.assertFalse(os.path.exists(os.path.join(ap.analysis_dir,'NewProj')))
+        # Import the project
+        import_project(ap,self.new_project_dir)
+        self.assertTrue('NewProj' in [p.name
+                                      for p in ap.get_analysis_projects()])
+        self.assertTrue('NewProj' in [p.name
+                                      for p in ap.get_analysis_projects_from_dirs()])
+        self.assertTrue(os.path.exists(os.path.join(ap.analysis_dir,'NewProj')))
+        # Verify via fresh AutoProcess object
+        ap2 = AutoProcess(mockdir.dirn)
+        self.assertTrue('NewProj' in [p.name
+                                      for p in ap2.get_analysis_projects()])
+        self.assertTrue('NewProj' in [p.name
+                                      for p in ap2.get_analysis_projects_from_dirs()])
+        self.assertTrue(os.path.exists(os.path.join(ap2.analysis_dir,'NewProj')))
+        # Check for QC report and ZIP file
+        print os.listdir(os.path.join(ap2.analysis_dir,'NewProj'))
+        for f in ("qc_report.html",
+                  "multiqc_report.html",
+                  "qc_report.NewProj.160621_M00879_0087_000000000-AGEW9_analysis.zip",):
+            f = os.path.join(ap2.analysis_dir,'NewProj',f)
+            self.assertTrue(os.path.exists(f),"Missing %s" % f)
 
     def test_import_project_already_in_metadata_file(self):
         """import_project: fails when project is already in projects.info
